refactor[#554] : SnackbarController index.ts 를 통해 import
