+++ src/shared/ui/modal/modal.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from "@storybook/react";
 import { OverlayPortal } from "@/app/OverlayPortal";
 import { useModal } from "@/shared/lib";
-import { SnackbarController } from "@/shared/store/snackbar";
+import { SnackbarController } from "@/shared/store";
 import { Modal } from "./modal";
 import { CenterModal, FullPageModal } from "./modal.mocks";
